+++ src/editorOptions.js
@@ -7,7 +7,7 @@ const editorOptions = [
   'enableBasicAutocompletion',
   'enableLiveAutocompletion',
   'enableSnippets',
-]
+];
 
 const editorEvents = [
   'onChange',
@@ -33,8 +33,16 @@ const getAceInstance = ()=>{
   }
   return ace;
 }
-export {
-  editorOptions,
-  editorEvents,
-  getAceInstance,
-}
+
+const debounce = (fn, delay) => {
+  var timer = null;
+  return function() {
+    var context = this,
+      args = arguments;
+    clearTimeout(timer);
+    timer = setTimeout(function() {
+      fn.apply(context, args);
+    }, delay);
+  };
+};
+export { editorOptions, editorEvents, debounce,getAceInstance };